requirements.txt
scipy
scikit-learn
numpy
opencv_python
Pillow
torch>=1.2.0
torchvision
tensorboardX
sentence-transformers
